addsigninandsignup

+++ src/actions/auth.actions.js
@@ -33,6 +33,11 @@ export const login = (user) => {
     }
 }
 
+
+
+
+
+
 export const isUserLoggedIn=()=>{
     return async (dispatch)=>{
         const token=localStorage.getItem('token')
@@ -57,11 +62,22 @@ export const isUserLoggedIn=()=>{
 }
 
 export const signout=()=>{
-   
+
     return async (dispatch)=>{
-        window.localStorage.clear();
-        dispatch({
-            type:authConstants.LOGOUT_REQUEST
-        })
+        dispatch({type:authConstants.LOGOUT_REQUEST})
+        const res=await axiosIntance.post("/admin/signout")
+        console.log(res)
+        if(res.status===200){
+            window.localStorage.clear();
+            dispatch({
+                type:authConstants.LOGOUT_SUCCESS
+            })
+        }else{
+                dispatch({
+                    type:authConstants.LOGOUT_FAILURE,
+                    payload:{error:res.data.error}
+                })
+        }
+     
     }
 }
